submit 1228 修改获取公告bug

+++ BackEnd/Administrator/src/main/java/com/football/administrator/Mapper/AnnouncementMapper.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Mapper
 public interface AnnouncementMapper {
     @Insert("INSERT INTO announcement (title,detail,time,receiver) VALUES (#{title},#{detail},#{time}, #{receiver})")
-    void postAnnouncement(String title, String detail, LocalDateTime time,Integer receiver);
+    void postAnnouncement(String title, String detail, LocalDateTime time,Long receiver);
 
     @Select("SELECT * FROM announcement WHERE receiver = #{id}")
     List<AnnouncementDTO> getAnnouncementList(Long id);
